+++ spec/savon/soap/xml_spec.rb
@@ -88,6 +88,17 @@ describe Savon::SOAP::XML do
     end
   end
 
+  describe "#env_namespace" do
+    it "should default to :env" do
+      xml.env_namespace.should == :env
+    end
+
+    it "should set the SOAP envelope namespace" do
+      xml.env_namespace = :soapenv
+      xml.env_namespace.should == :soapenv
+    end
+  end
+
   describe "#namespaces" do
     it "should default to a Hash containing the namespace for SOAP 1.1" do
       xml.namespaces.should == { "xmlns:env" => "http://schemas.xmlsoap.org/soap/envelope/" }
@@ -141,6 +152,10 @@ describe Savon::SOAP::XML do
     context "by default" do
       it "should start with an XML declaration" do
         xml.to_xml.should match(/^<\?xml version="1.0" encoding="UTF-8"\?>/)
+      end
+
+      it "should use default SOAP envelope namespace" do
+        xml.to_xml.should include("<env:Envelope", "<env:Body")
       end
 
       it "should add the xsd namespace" do
@@ -201,6 +216,13 @@ describe Savon::SOAP::XML do
         uri = "http://schemas.xmlsoap.org/soap/envelope/"
         xml.to_xml.should match(/<env:Envelope (.*)xmlns:env="#{uri}"(.*)>/)
         reset_soap_version
+      end
+    end
+
+    context "with the SOAP envelope namespace set to an empty String" do
+      it "should not add a namespace to SOAP envelope tags" do
+        xml.env_namespace = ""
+        xml.to_xml.should include("<Envelope", "<Body")
       end
     end